Common implementation for all hash types
Summary:
Code was very duplicated between ContentId, Sha1, Sha256. And in this stack I plan to implement even more hashes, so I made this macro which makes it really easy to do so.

I took the opportunity to make the inner field not public, and only accessible via from/into.

Reviewed By: liubov-dmitrieva

Differential Revision: D29992279

fbshipit-source-id: c0b7225a3634071a1b1513119ec516d14bd8fd9e