Create simple bonsai commit
Summary:
On `createremote`, create a changeset using the create_bonsai_changesets method created earlier.

For now, this changeset is create with a bunch of placeholders, but I added todo for all of those things and will tackle them over the next diffs, otherwise this would be a massive diff and hard to review.

Reviewed By: liubov-dmitrieva

Differential Revision: D29990295

fbshipit-source-id: 6b4c97887c0b4c017c586bf0ea06f12df9d07d23