checkout: add parallelization to checkout scmstore path
Summary:
Same as the previous diff, but for the scmstore path.

Previously file fetching during native checkout wasn't done in parallel, so if
the fetcher was stuck downloading one large file, other files would not be
downloaded while it waited. This diff breaks the file fetching into smaller
chunks and has the chunks fetched in parallel.

Reviewed By: andll

Differential Revision: D30075316

fbshipit-source-id: 42c0401212373296c0ccbf65a676dec80a8e7a90