Add ChangesetId to UploadToken and use it on /upload/changeset/bonsai
Summary:
Using the new macros from previous diffs, this creates a new `ChangesetId` edenapi type and adds it to AnyId, which allows it to be used from UploadToken.

It then adds the lookup method for it, and returns it from upload_bonsai_changeset call (instead of a fake HgId UploadToken).

This will be used so that the client can know the changeset id of the uploaded snapshot.

Reviewed By: StanislavGlebik

Differential Revision: D29995429

fbshipit-source-id: e2ee4b9b0ac21d6f5394afacbfed1802da64013b