fastlog: disable fastlog if graphql client cannot be constructed
Summary:
The "local log" algorithm in the fastlog extension is poorly implemented.
It can be much slower than the non-fastlog algorithm. This means, if the
"remote" fastlog thread does not work in production, then enabling fastlog
can actually seriously regress log performance.

The fastlog test (already) cannot meaningfully test the remote behavior. So it
is removed for now.

Differential Revision: D30145651

fbshipit-source-id: 601a38e03920fbcae86e23262949ae3e3a16b9af