dag: add failpoints for not found errors
Summary:
Similar to D30094157. This is helpful to get a Rust backtrace on "not found"
errors. This is useful because the Python traceback might not have Rust
stack.

Ideally, it'd be better if we can integrate backtrace into the error types.
However, the stdlib backtrace is still unstable and the thirdparty backtrace
seems too heavyweight.

Reviewed By: DurhamG

Differential Revision: D30123355

fbshipit-source-id: 7af76422143340194d8499a75e8ed65e9bf1fcae