Improve ODS logging on unbundle
Summary:
The time between logging total_unbundles and the actual successes is the time of the unbundle operation, which may be long.

This makes the alarm on D30222804 much less accurate, as success and fail for the same operation might fall in different buckets.

This diff changes two things:
- total_unbundles are logged at the end of the unbundle operation, which should make tracking more accurate when compared against successes.
- resolver_error is now logged in more cases that would previously error but not be logged

I created a wrapper function in order to make sure it always logs, as before it could not log if there were some early errors.

Differential Revision: D30248117

fbshipit-source-id: 4ec0c148dd7aa818b6d204fafecacacf4d267be7