support streaming for file uploads rather than buffering the whole content
Summary:
support streaming for file uploads rather than buffering the whole content

This is a preferable way for big files. We are currently using LFS endpoint instead for big files but not in snapshots.

Also, it we enable streaming, we have an option not to use LFS endpoint for file uploads in the future.

**[Land after hg release with D30100887 (https://github.com/facebookexperimental/eden/commit/fe8ed9d28c23892a6d920a72c92fd54b2da641b6) has been fully rolled out]**

Reviewed By: yancouto

Differential Revision: D30158390

fbshipit-source-id: b62c498b8bdf23a5f413f6e4b71d7433906e4611