mononoke: store hashed paths in mutable_renames
Summary:
Mysql has a limit on the length of the [key used in the index](https://dev.mysql.com/doc/refman/8.0/en/innodb-limits.html#:~:text=The%20index%20key%20prefix%20length,REDUNDANT%20or%20COMPACT%20row%20format.)

So we can't use the full path in index as-is. We can try to use just prefix of
the path in the index, but that won't allow us to make [the index
unique](https://stackoverflow.com/questions/15157227/mysql-varchar-index-length).

So let's do the same thing we do in filenodes - hash the path and store hash ->
path mapping in the separate table. Because we do similar thing in filenodes it
we can reuse some of the code we used there.

Reviewed By: markbt

Differential Revision: D30301729

fbshipit-source-id: 32c058a163e5ff541641c6049a74168ceba66a74