test_support: canonicalize the temporary directory path
Summary:
With D30320515, EdenFS internally canonicalize all the mount paths passed to
it. As a result, the output of `eden list` may not match the path given to
`eden mount` if one of the directory leading to the mount point was a symlink.
Since some tests are comparing both, this can lead to some test failures.

To solve this, we simply need to make sure that the temporary directory for the
test is canonicalized.

Reviewed By: fanzeyi

Differential Revision: D30349411

fbshipit-source-id: 139d4be02b5783c6a439270845239acab6a6c955