mononoke: use mutable_renames in fastlog
Summary:
Let's make it possible to query mutable renames from fastlog. For now this is a
very basic support i.e. we don't support middle-of-history renames, blame is
not supported etc.

Mutable rename logic is gated by a tunable, so we can roll it back quickly in
case of problems.

Reviewed By: ahornby

Differential Revision: D30279932

fbshipit-source-id: 0e8e329e8ab4d4980ab401bd103e6c97419d0f67