derived_data: introduce DerivedDataManager
Summary:
The `DerivedDataManager` will manage the ordering of derivation for derived
data, taking into account dependencies between types as well as the topological
ordering of the repository.  It will replace the free functions in
`derived_data` as well as much of the `utils` crate.

This is the first step: it introduces the manager, although currently it only takes
over management of the derived data lease.

Reviewed By: mitrandir77

Differential Revision: D30281634

fbshipit-source-id: 04c3a34d97ea02cc8c26d34096cca341e800da9b