mononoke: add a binary that can sync changesets to hg servers
Summary:
Simple binary that can be used for syncing changesets to hg servers. It's very
simple - it just moves a bookmark, and then wait until hg sync job syncs it
(obviously it means that if sync job is not running, then it isn't going to be
synced).

It also supports syncing only a straight line of commits with no merges -
merges add complexity here, so I decided to not deal with this complexity for
now.

Reviewed By: mitrandir77

Differential Revision: D30447234

fbshipit-source-id: e4624586e4fc53212c1b13a2cd622aa9474a20b8