mononoke: remove history variable from visit function
Summary:
There were two variables that controlled the output of `list_file_history` -
one was `history` variable, another was `bfs`. `history` variable got all
immediate ancestors, and bfs was used to fetch new ancestors.

This split makes it hard to add generation number ordering, so in this diff I
suggest to remove `history` variable altogether and just use `bfs` to control
the order of the history.

Reviewed By: mitrandir77

Differential Revision: D30483094

fbshipit-source-id: 0a4cac771383e17e61f58354a30d4e6db7e6547f