mononoke: add ordering by gen num in fastlog
Summary:
To be precise, it's ordering by generation number and parent order i.e. we'd
like to show first parents ahead of second parents.
Note that parent ordering is very basic at the moment, and won't always order
commits correctly when both parents have the same generation numbers. We can
improve it in the future, but I believe it shouldn't be a big issue now.

Reviewed By: mitrandir77

Differential Revision: D30483089

fbshipit-source-id: 67e13b5757831d652b57d6ad42b6135005a0b621