minor smartservice updates for consistency with predictive fetch thrift endpoint in daemon
Summary:
Made changes to ensure that numResults is always a 32 bit unsigned int, and startTime and endTime are 64 bit unsigned ints. This is to ensure consistency across the smartservice and the endpoint in the daemon.

Also, updated the scuba query in the smartservice to only consider dirs with > 1 access (may update this later to accept a configurable lower bound on access count, but for now, including access=1 doesn't make sense).

Reviewed By: genevievehelsel

Differential Revision: D30396526

fbshipit-source-id: 10e7bd969928da91ab29d413280a1ff956db438c