treemanifest: change diff's matcher to not hold the gil
Summary:
In a future diff we want to start allowing a background thread to fetch
tree data while the diff algorithm is running. To do so, we need the gil to not
be held during the diff algorithm. But, the diff algorithm needs to run the
matcher, and the matcher may be in Python. So let's introduce a new matcher
wrapper that acquires the gil.

In the future we should get rid of the python matchers entirely, which will
eliminate this problem.

Reviewed By: kulshrax

Differential Revision: D29971826

fbshipit-source-id: 8a9ba0ea65a0b4748e39178cdf4a08c922755b02