treemanifest: moving diff fetching to a background thread
Summary:
When testing hg checkout with an empty cache, I noticed the tree
diff'ing phase would alternate between downloading data and doing cpu work. This
implied it was all synchronous. This diff speeds it up by doing the downloading
on a background thread, while the main thread does the actual diff.

Ideally we'd make the actual diff itself parallel as well, but the tree
structure is a tree of semi-mutable references, which would require a larger
refactoring.

Reviewed By: quark-zju

Differential Revision: D29971824

fbshipit-source-id: b4d12a0de8af8e18e8a420d43088663a795db5f8