manifest: convert Link to contain an Arc
Summary:
We want to make tree traversals parallelized, but to do that it needs
to be safe to hand different parts of the tree to different threads. The current
structure involves the tree being a straight forward tree of
`Link->BTreeMap<Path, Link>` and the algorithms largely act on references to
Link's. This diff converts Link to actually contain an inner Arc, so in later
diffs we can stop using references and then start sending trees across threads.

Note: The change affects the mutability pattern for trees in an unusual way. See
code comments for details.

Note: The inner-Arc pattern was used (instead of wrapping Link's in Arc) because
generally don't want the Link's to be clonable and we want to avoid the
accidental clones that Arc allows.

Reviewed By: andll

Differential Revision: D30132761

fbshipit-source-id: 8a47b88fc27993420af3d486cbf4e80209cdb3ac