treemanifest: use Rust store's directly in treemanifest
Summary:
Previously treemanifest kept a reference to the Python store, and any
call would have to go up to Python then come back down to Rust. Since our stores
are now 100% Rust (at least at the top layer), let's just downcast it to the
appropriate Rust store and store that instead.

This helps in a future diff where we want to access the store without taking the
gil.

Reviewed By: kulshrax

Differential Revision: D29971828

fbshipit-source-id: 77ff11897045282c9e6a6029b126dcdd20c8e9db