snapshot: Handle files that were deleted but still exist
Summary:
This fixes a small bug in snapshots, as noted in the `TODO` comment I deleted.

If you `hg rm` a file and then replace it with something else, we want that to be uploaded to the snapshot as well.

These types of files are uploaded as "UntrackedChange", which is a bit confusing but correct, it means the file was deleted and now there are untracked changes in it (similar to how if a file is marked as "untracked deletion" but it didn't exist, it means it was `hg add`ed and then locally deleted)

Differential Revision: D30398143

fbshipit-source-id: dbd316fbcdc0e9b1781c5c967eb12c02805f4361