snapshot: Upload file content to ephemeral blobstore
Summary:
The snapshot itself is uploaded to ephemeral blobstore, but before this diff, file content was still uploaded to the persistent one.

Server side logic to upload to bubble was already added some time ago, but we needed the client part of it.

There is a slight bug, since the logic to upload files still checks if the file was already uploaded in the **persistent** blobstore, so if it's there it won't be uploaded again to the ephemeral blobstore. This is fixed on the next diff (D30396437).

Differential Revision: D30396432

fbshipit-source-id: e6e5ec299ec5323503eedba32956569e16f3eb50