Upload snapshot to ephemeral blobstore
Summary:
This diff makes snapshots be uploaded to the ephemeral blobstore.

Given the previous diffs which implemented a "repo view" and piped bubble id to edenapi server, on this diff I do some BlobRepo refactoring which allows CreateChangeset to use repo view instead of BlobRepo to save the snapshot changeset.

Reviewed By: StanislavGlebik

Differential Revision: D30370980

fbshipit-source-id: b1a6a7d9d061dd09d6973c0a6942c22e86387f4a