edenapi: Allow upload tokens for bubbles
Summary:
Objects (for now, files and changesets) can be in the persistent or ephemeral blobstore (in any of the bubbles). This diff makes it so `UploadToken`s can reflect that.

Nicely, for objects in persistent blobstore, this shouldn't affect the size of anything shipped on the wire, as its serialisation gets skipped.

 ---

An unrelated but similar problem I noticed is that `UploadToken`s don't capture the repo the object was uploaded to. So it is theoretically possible to upload something in one repo and "prove" to mononoke that it was uploaded to another repo (well, once signatures are actually used). This seems very hard to actually do in practice without bad intent (or even with it), so I'm not sure if it's something we want to fix.

Reviewed By: StanislavGlebik

Differential Revision: D30396437

fbshipit-source-id: e278cd73061d3aaf7e3353236503d239a366ed5d