edenapi: Make python-aware rust code minimal for uploadsnapshot
Summary:
This diff starts making the python binding layer simpler. It's better to make the python-aware Rust code as simple as possible, to make reusability easier.

This diff moves the `uploadsnapshot` functionality to a "pure-rust" file, and makes it in an easily extensible way, so that when adding more functionality we don't end up with a huge file with everything.

Reviewed By: ahornby

Differential Revision: D30429524

fbshipit-source-id: e46e71a8993917720ab38b50a8c967564ac82fb6