mononoke: add retries in add_sync_target derived data derivations
Summary:
add_sync_target might need to derive a lot of data, and it takes a long time to
do it. We don't have any resumability, so if it fails for any reason, then we'd
need to start over.

For now let's just retry a few times so that we don't have to start over
because of flakiness

Reviewed By: mitrandir77

Differential Revision: D30511785

fbshipit-source-id: 1a9c5e62db366022ad487ed108dd41b1dea4caa2