sparse: add strongly typed sparse config types
Summary:
Previously the SparseConfig type was abused to mean three things:
1. The non-recursive contents of .hg/sparse
2. The non-recursive contents of a sparse profile.
3. The recursively expanded contents of enabled sparse rules and profiles.

In a later diff we'll be changing how includes/excludes work to make it more
flexible, and this reuse of the same type makes it complicated.

So let's split it into three distinct types. Later diffs will change them so
they aren't so similar. In doing so, we also slightly change the way sparse
profiles are loaded, to load them recursively.

Reviewed By: andll

Differential Revision: D30284701

fbshipit-source-id: f908ba167a04f9863ab75e1cd6a94027fa64607e