add a new function to handle history fixup deletes
Summary:
In one of our large repos the files were synced using ad-hoc scripts between
the repos before the megarepo. This made the history and blame full of
irrelevant commit messages like "sync from repo x". The new functionality will
allow us to re-merge those files with correct histories. To do that we need to
delete the:
 * files with wrong history from master branch
 * all other files from the source branch with correct history - so only those
   are merged in

As an optimisation we're not removing the files that are the same on both
branches from branch with correct history. This will make the merge bonsai
smaller while keeping the history the same (at least when unode v2s are
enabled).

Reviewed By: StanislavGlebik

Differential Revision: D30456465

fbshipit-source-id: f14e8caf8e104c8d8069edc098cfe6419bbe407e