treemanifest: moving diff fetching to a background thread
Summary:
When testing hg checkout with an empty cache, I noticed the tree
diff'ing phase would alternate between downloading data and doing cpu work. This
implied it was all synchronous. This diff speeds it up by doing the downloading
on a background thread, while the main thread does the actual diff.

Ideally we'd make the actual diff itself parallel as well, but the tree
structure is a tree of semi-mutable references, which would require a larger
refactoring.

Reviewed By: andll

Differential Revision: D30553728

fbshipit-source-id: 4ea953909827cf1ec4fd67ac297f63bfadd67483