sparse: support hg sparse files --rev $REV
Summary:
I need to be able to run 'hg sparse files $PROFILE' in an Eden checkout
for a Sandcastle job that analyzes sparse profiles. We can't run 'hg sparse' at
all in Eden repos, so instead I run it in the backing repo. But that repo is
checked out to the null commit, so let's support --rev on 'hg sparse files' so I
can inspect an arbitrary rev.

Differential Revision: D30561529

fbshipit-source-id: 93b46caa9b63637bf4d63d5438bd23cbffb3983a